Keystone icehouse-3 milestone (2014.1.b3)

+++ keystone/assignment/core.py
@@ -1,5 +1,3 @@
-# vim: tabstop=4 shiftwidth=4 softtabstop=4
-
 # Copyright 2012 OpenStack Foundation
 #
 # Licensed under the Apache License, Version 2.0 (the "License"); you may
@@ -23,6 +21,7 @@ import six
 from keystone import clean
 from keystone.common import cache
 from keystone.common import dependency
+from keystone.common import driver_hints
 from keystone.common import manager
 from keystone import config
 from keystone import exception
@@ -34,15 +33,21 @@ CONF = config.CONF
 LOG = log.getLogger(__name__)
 SHOULD_CACHE = cache.should_cache_fn('assignment')
 
-DEFAULT_DOMAIN = {'description':
-                  (u'Owns users and tenants (i.e. projects)'
-                   ' available on Identity API v2.'),
-                  'enabled': True,
-                  'id': CONF.identity.default_domain_id,
-                  'name': u'Default'}
+# NOTE(blk-u): The config option is not available at import time.
+EXPIRATION_TIME = lambda: CONF.assignment.cache_time
+
+
+def calc_default_domain():
+    return {'description':
+            (u'Owns users and tenants (i.e. projects)'
+                ' available on Identity API v2.'),
+            'enabled': True,
+            'id': CONF.identity.default_domain_id,
+            'name': u'Default'}
 
 
 @dependency.provider('assignment_api')
+@dependency.optional('revoke_api')
 @dependency.requires('credential_api', 'identity_api', 'token_api')
 class Manager(manager.Manager):
     """Default pivot point for the Assignment backend.
@@ -53,6 +58,7 @@ class Manager(manager.Manager):
     The late import works around this.  The if block prevents creation of the
     api object by both managers.
     """
+    _PROJECT = 'project'
 
     def __init__(self):
         assignment_driver = CONF.assignment.driver
@@ -63,7 +69,7 @@ class Manager(manager.Manager):
 
         super(Manager, self).__init__(assignment_driver)
 
-    @notifications.created('project')
+    @notifications.created(_PROJECT)
     def create_project(self, tenant_id, tenant):
         tenant = tenant.copy()
         tenant.setdefault('enabled', True)
@@ -76,22 +82,26 @@ class Manager(manager.Manager):
                                          ret['domain_id'])
         return ret
 
-    @notifications.updated('project')
+    @notifications.disabled(_PROJECT, public=False)
+    def _disable_project(self, tenant_id):
+        return self.token_api.delete_tokens_for_users(
+            self.list_user_ids_for_project(tenant_id),
+            project_id=tenant_id)
+
+    @notifications.updated(_PROJECT)
     def update_project(self, tenant_id, tenant):
         tenant = tenant.copy()
         if 'enabled' in tenant:
             tenant['enabled'] = clean.project_enabled(tenant['enabled'])
         if not tenant.get('enabled', True):
-            self.token_api.delete_tokens_for_users(
-                self.list_user_ids_for_project(tenant_id),
-                project_id=tenant_id)
+            self._disable_project(tenant_id)
         ret = self.driver.update_project(tenant_id, tenant)
         self.get_project.invalidate(self, tenant_id)
         self.get_project_by_name.invalidate(self, ret['name'],
                                             ret['domain_id'])
         return ret
 
-    @notifications.deleted('project')
+    @notifications.deleted(_PROJECT)
     def delete_project(self, tenant_id):
         project = self.driver.get_project(tenant_id)
         user_ids = self.list_user_ids_for_project(tenant_id)
@@ -255,12 +265,18 @@ class Manager(manager.Manager):
                 self.driver.remove_role_from_user_and_project(user_id,
                                                               tenant_id,
                                                               role_id)
+                if self.revoke_api:
+                    self.revoke_api.revoke_by_grant(role_id, user_id=user_id,
+                                                    project_id=tenant_id)
+
             except exception.RoleNotFound:
                 LOG.debug(_("Removing role %s failed because it does not "
                             "exist."),
                           role_id)
 
-    def list_projects_for_user(self, user_id):
+    # TODO(henry-nash): We might want to consider list limiting this at some
+    # point in the future.
+    def list_projects_for_user(self, user_id, hints=None):
         # NOTE(henry-nash): In order to get a complete list of user projects,
         # the driver will need to look at group assignments.  To avoid cross
         # calling between the assignment and identity driver we get the group
@@ -270,18 +286,20 @@ class Manager(manager.Manager):
 
         group_ids = [x['id'] for
                      x in self.identity_api.list_groups_for_user(user_id)]
-        return self.driver.list_projects_for_user(user_id, group_ids)
+        return self.driver.list_projects_for_user(
+            user_id, group_ids, hints or driver_hints.Hints())
 
     @cache.on_arguments(should_cache_fn=SHOULD_CACHE,
-                        expiration_time=CONF.assignment.cache_time)
+                        expiration_time=EXPIRATION_TIME)
     def get_domain(self, domain_id):
         return self.driver.get_domain(domain_id)
 
     @cache.on_arguments(should_cache_fn=SHOULD_CACHE,
-                        expiration_time=CONF.assignment.cache_time)
+                        expiration_time=EXPIRATION_TIME)
     def get_domain_by_name(self, domain_name):
         return self.driver.get_domain_by_name(domain_name)
 
+    @notifications.created('domain')
     def create_domain(self, domain_id, domain):
         ret = self.driver.create_domain(domain_id, domain)
         if SHOULD_CACHE(ret):
@@ -289,16 +307,26 @@ class Manager(manager.Manager):
             self.get_domain_by_name.set(ret, self, ret['name'])
         return ret
 
+    @manager.response_truncated
+    def list_domains(self, hints=None):
+        return self.driver.list_domains(hints or driver_hints.Hints())
+
+    @notifications.disabled('domain', public=False)
+    def _disable_domain(self, domain_id):
+        self.token_api.delete_tokens_for_domain(domain_id)
+
+    @notifications.updated('domain')
     def update_domain(self, domain_id, domain):
         ret = self.driver.update_domain(domain_id, domain)
         # disable owned users & projects when the API user specifically set
         #     enabled=False
         if not domain.get('enabled', True):
-            self.token_api.delete_tokens_for_domain(domain_id)
+            self._disable_domain(domain_id)
         self.get_domain.invalidate(self, domain_id)
         self.get_domain_by_name.invalidate(self, ret['name'])
         return ret
 
+    @notifications.deleted('domain')
     def delete_domain(self, domain_id):
         # explicitly forbid deleting the default domain (this should be a
         # carefully orchestrated manual process involving configuration
@@ -315,7 +343,8 @@ class Manager(manager.Manager):
         # to get a valid token to issue this delete.
         if domain['enabled']:
             raise exception.ForbiddenAction(
-                action=_('delete a domain that is not disabled'))
+                action=_('cannot delete a domain that is enabled, '
+                         'please disable it first.'))
 
         self._delete_domain_contents(domain_id)
         self.driver.delete_domain(domain_id)
@@ -382,18 +411,32 @@ class Manager(manager.Manager):
                               {'userid': user['id'],
                                'domainid': domain_id})
 
+    @manager.response_truncated
+    def list_projects(self, hints=None):
+        return self.driver.list_projects(hints or driver_hints.Hints())
+
+    # NOTE(henry-nash): list_projects_in_domain is actually an internal method
+    # and not exposed via the API.  Therefore there is no need to support
+    # driver hints for it.
+    def list_projects_in_domain(self, domain_id):
+        return self.driver.list_projects_in_domain(domain_id)
+
+    def list_user_projects(self, user_id, hints=None):
+        return self.driver.list_user_projects(
+            user_id, hints or driver_hints.Hints())
+
     @cache.on_arguments(should_cache_fn=SHOULD_CACHE,
-                        expiration_time=CONF.assignment.cache_time)
+                        expiration_time=EXPIRATION_TIME)
     def get_project(self, project_id):
         return self.driver.get_project(project_id)
 
     @cache.on_arguments(should_cache_fn=SHOULD_CACHE,
-                        expiration_time=CONF.assignment.cache_time)
+                        expiration_time=EXPIRATION_TIME)
     def get_project_by_name(self, tenant_name, domain_id):
         return self.driver.get_project_by_name(tenant_name, domain_id)
 
     @cache.on_arguments(should_cache_fn=SHOULD_CACHE,
-                        expiration_time=CONF.assignment.cache_time)
+                        expiration_time=EXPIRATION_TIME)
     def get_role(self, role_id):
         return self.driver.get_role(role_id)
 
@@ -403,6 +446,10 @@ class Manager(manager.Manager):
         if SHOULD_CACHE(ret):
             self.get_role.set(ret, self, role_id)
         return ret
+
+    @manager.response_truncated
+    def list_roles(self, hints=None):
+        return self.driver.list_roles(hints or driver_hints.Hints())
 
     @notifications.updated('role')
     def update_role(self, role_id, role):
@@ -441,20 +488,34 @@ class Manager(manager.Manager):
     def remove_role_from_user_and_project(self, user_id, tenant_id, role_id):
         self.driver.remove_role_from_user_and_project(user_id, tenant_id,
                                                       role_id)
-        self.token_api.delete_tokens_for_user(user_id)
+        if CONF.token.revoke_by_id:
+            self.token_api.delete_tokens_for_user(user_id)
+        if self.revoke_api:
+            self.revoke_api.revoke_by_grant(role_id, user_id=user_id,
+                                            project_id=tenant_id)
 
     def delete_grant(self, role_id, user_id=None, group_id=None,
                      domain_id=None, project_id=None,
                      inherited_to_projects=False):
         user_ids = []
-        if group_id is not None:
-            # NOTE(morganfainberg): The user ids are the important part for
-            # invalidating tokens below, so extract them here.
+        if group_id is None:
+            if self.revoke_api:
+                self.revoke_api.revoke_by_grant(user_id=user_id,
+                                                role_id=role_id,
+                                                domain_id=domain_id,
+                                                project_id=project_id)
+        else:
             try:
+                # NOTE(morganfainberg): The user ids are the important part
+                # for invalidating tokens below, so extract them here.
                 for user in self.identity_api.list_users_in_group(group_id,
                                                                   domain_id):
                     if user['id'] != user_id:
                         user_ids.append(user['id'])
+                        if self.revoke_api:
+                            self.revoke_api.revoke_by_grant(
+                                user_id=user['id'], role_id=role_id,
+                                domain_id=domain_id, project_id=project_id)
             except exception.GroupNotFound:
                 LOG.debug(_('Group %s not found, no tokens to invalidate.'),
                           group_id)
@@ -563,6 +624,9 @@ class Driver(object):
         role_set.remove(frozenset(self._role_to_dict(role_id,
                                                      inherited).items()))
         return [dict(r) for r in role_set]
+
+    def _get_list_limit(self):
+        return CONF.assignment.list_limit or CONF.list_limit
 
     @abc.abstractmethod
     def get_project_by_name(self, tenant_name, domain_id):
@@ -683,8 +747,11 @@ class Driver(object):
         raise exception.NotImplemented()
 
     @abc.abstractmethod
-    def list_domains(self):
-        """List all domains in the system.
+    def list_domains(self, hints):
+        """List domains in the system.
+
+        :param hints: filter hints which the driver should
+                      implement if at all possible.
 
         :returns: a list of domain_refs or an empty list.
 
@@ -741,8 +808,11 @@ class Driver(object):
         raise exception.NotImplemented()
 
     @abc.abstractmethod
-    def list_projects(self, domain_id=None):
-        """List all projects in the system.
+    def list_projects(self, hints):
+        """List projects in the system.
+
+        :param hints: filter hints which the driver should
+                      implement if at all possible.
 
         :returns: a list of project_refs or an empty list.
 
@@ -750,15 +820,70 @@ class Driver(object):
         raise exception.NotImplemented()
 
     @abc.abstractmethod
-    def list_projects_for_user(self, user_id, group_ids):
+    def list_projects_in_domain(self, domain_id):
+        """List projects in the domain.
+
+        :param domain_id: the driver MUST only return projects
+                          within this domain.
+
+        :returns: a list of project_refs or an empty list.
+
+        """
+        raise exception.NotImplemented()
+
+    @abc.abstractmethod
+    def list_projects_for_user(self, user_id, group_ids, hints):
         """List all projects associated with a given user.
 
         :param user_id: the user in question
         :param group_ids: the groups this user is a member of.  This list is
                           built in the Manager, so that the driver itself
                           does not have to call across to identity.
+        :param hints: filter hints which the driver should
+                      implement if at all possible.
 
         :returns: a list of project_refs or an empty list.
+
+        """
+        raise exception.NotImplemented()
+
+    @abc.abstractmethod
+    def get_roles_for_groups(self, group_ids, project_id=None, domain_id=None):
+        """List all the roles assigned to groups on either domain or
+        project.
+
+        If the project_id is not None, this value will be used, no matter what
+        was specified in the domain_id.
+
+        :param group_ids: iterable with group ids
+        :param project_id: id of the project
+        :param domain_id: id of the domain
+
+        :raises: AttributeError: In case both project_id and domain_id are set
+                                 to None
+
+        :returns: a list of Role entities matching groups and
+                  project_id or domain_id
+
+        """
+        raise exception.NotImplemented()
+
+    @abc.abstractmethod
+    def list_projects_for_groups(self, group_ids):
+        """List projects accessible to specified groups.
+
+        :param group_ids: List of group ids.
+        :returns: List of projects accessible to specified groups.
+
+        """
+        raise exception.NotImplemented()
+
+    @abc.abstractmethod
+    def list_domains_for_groups(self, group_ids):
+        """List domains accessible to specified groups.
+
+        :param group_ids: List of group ids.
+        :returns: List of domains accessible to specified groups.
 
         """
         raise exception.NotImplemented()
@@ -792,7 +917,6 @@ class Driver(object):
         """
         raise exception.NotImplemented()
 
-    """Interface description for an assignment driver."""
     # role crud
 
     @abc.abstractmethod
@@ -805,8 +929,11 @@ class Driver(object):
         raise exception.NotImplemented()
 
     @abc.abstractmethod
-    def list_roles(self):
-        """List all roles in the system.
+    def list_roles(self, hints):
+        """List roles in the system.
+
+        :param hints: filter hints which the driver should
+                      implement if at all possible.
 
         :returns: a list of role_refs or an empty list.
